IDS/backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-filter==24.2
